requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
aiomysql==0.2.0
python-multipart==0.0.19
websockets==14.1
psutil==6.1.1
aiohttp==3.11.11
aiofiles==24.1.0
bcrypt==4.2.1
pyjwt==2.10.1
jinja2==3.1.4
cryptography==44.0.0
certifi>=2024.0.0
paramiko>=3.0.0
